added ul, created validate function, added ul function
I changed the CSS to only target the <p> tag with the error id instead of every <p> tag. I created the validate function, though it isn't working yet. I made the code that adds a list element to the <ul> on the page. I changed the form method and removed the names of all the inputs since they seemed to be redundant. But the form and associated code are both not working, and it seems like the page refreshes once I hit submit on the form.