List creation functions properly, added success notification
I misunderstood the task to mean that the list creation code should run when the javascript loads. It is meant to add the name of a task that is submitted, and so I moved the code to within the validate function. I also added a success message that displays when a value is successfully submitted. Made the text in both success and error messages center aligned. Changed the error message text slightly to make it feel more impassive.